requirements.txt
multiprocessing
opencv-python3
RPi.GPIO